refactor: :recycle: Move the validators into their respective fields
