+++ src/types/interfaces/Validator.ts
@@ -1,0 +1,4 @@
+export default interface Validator {
+    isValid: (value: string) => boolean;
+    error: string;
+}
